Support configurable resolvers
The default import resolution strategy is to use `require.resolve`.
Custom import resolution strategies can be passed in via

  - `--import-resolver` (cli)
  - as `resolveImport` (lib).
  - as the option `importResolver` (eslint)

For the CLI and eslint, the resolver must be a string that can be
`require`d, and whose export is a resolver function.

For the LIB `resolveImport` must be a resolver function.

A resolver function implements

  (
    identifier: string,
    options: { basedir: string, ...args: [any]}
  ) => string

Returning the absolute path to the resolved import, or throwing an error
e with `e.code === MODULE_NOT_FOUND` otherwise.